cleanup: Clean up Shader Editor panel
