+++ panels/__init__.py
@@ -9,7 +9,6 @@ submodules = [
     "extras_panels",
     "layers_panels",
     "quick_tools_panels",
-    "shader_editor_panels",
 ]
 
 register, unregister = register_submodule_factory(__name__, submodules)
